make bot more responsive and interesting
Add "random topic of the day" feature.  It will read these
from topics.txt if it exists in the current directory, and choose
a line at random based on the calendar date.

Also, respond to some messages without a wakeword or @-mention,
when we have posted to the channel in the last 2 minutes and
at least one of the following is true:
 - the new message ends with a ? or !
 - we randomly choose to, at a 20% chance of replying