attempt at fixing conversation loops
Look to see if the bot is making the same statement repeatedly.  If so,
truncate the message history it gets to elimiate the last repeated
response and everything before it.

This won't work on loops more than 1 sentence long, but it's a start.

Also, change the unsolicted response timer to only be reset when
the bot is mentioned.  Previously it would reset its timer even if
the previous response was also unsolicted.  This means it could keep
on going for a long time without prompts.  Now, it will require
occasional prompts to continue the conversation.