add "reply in thread" feature
When summoned in a public channel, the bot will now create
a public thread to reply on.

By summoning, I mean an @mention or using one of the bot's
wakewors.

For this to work, the user summoning the bot must have
"Create Public Threads" permission.  If the user doesn't,
then the bot will not reply.

I considered having the bot just reply in the main channel
in this case, but that creates a lot of strange edge-cases
about what happens when someone who DOES have thread-create
permission follows up with a comment in that same channel.
Will a separate thread then be created?  If so you'd have
the conversation split between channel and thread.  It also
could end up with multiple threads being created for the
same conversation.

In all, it's simpler to just not have the bot respond in
that case.