include thread-starting message in its history
When returning the history of a thread, Discord
does not include the message that kicked off the thread.

It will show it in the UI as if it were, but it's not
one of the messages returned by the history iterator.

This change adds that message into the history, if we
have room for it.