add /lobotomize bot command
This will cause the bot to "forget" all messages
that have been sent in the past, in the channel
the command is run from.

This might be useful if the bot enters a loop
of repeating messages, or if it keeps referencing
an old topic of conversation.

By default, any user can run this command, but
it can be limited by admins within the Discord
server UI.