fixes for "don't split responses" mode, decrease token budget
When the bot doesn't split responses into multiple messages,
the average length of each messages increases by a lot.  This
means that it's easy to run out of tokens.  So, I increased
the token cost when the bot is in "don't split responses" mode.

I also noticed that with my model (a llama 13b version), the
actual token budget available was around 740, not 2048.  So,
I decreased the default budget to this.

With these combined, if you run the bot with the
--dont-split-responses option, the earlier token budget
calculations would now realize you're over-budget when launching
the app, and so the app wouldn't launch.  To make this simpler
for folks in the future, I decreased the default token number
of history lines down to 7 (from 15).

Users who don't use the --dont-split-responses option
can incrase this back up to 15 if they want.