Show user an error message when image gen fails
when a user tries to generate an image, if the image generation fails,
the bot will now send an error message to the user.

This message is a template, so can be customized once template
customization is implemented.

Also, move all template-related code into templates.py and use
the google-style-sanctioned import style to import it.