fix user nicknames showing up in bot prompt
Previously, the bot prompt would show the user's discord account
name, even if they had a nickname set on the server. This commit
fixes that.

Turns out that all we needed to do was turn on the `members`
intent on the client.  Conveniently, we already had been
granted this on the server.