add /say command
/say forces the bot to say the passed-in text.
Other users in the room will see who used say,
and the bot will say the text.

/say is available to anyone with slash-command permission
for the bot, so you might want to restrict that.

when --reply-in-thread is enabled, the /say command
will not work in public channels.  It will still work
in other contexts, such as DMs, group chats, and of
course threads.

Also, refactor common discord-related code into
discord_utils.py