fix https://github.com/chrisrude/oobabot/issues/20
Oobabot wasn't running on python < 3.10 because it used
fancy new typing syntax.  This commit fixes that, returning
to syntax supported in python 3.8+.

I also added a new test to make sure that the bot can
at least be instantiated without crashing.  This would
have caught the issue earlier.

To do this, I refactored the oobabot file itself to be
a class, which actually cleaned it up nicely.  As part
of this I moved connection testing to the http_client
class, which is a good idea anyway.