Fix https://github.com/chrisrude/oobabot/issues/16 for real
Discord demands a response in <5 seconds, so we need to
defer() the interaction before regenerating the image.

This means the "try again" button will appear enabled
while the operation is in process, so instead disable
all the buttons until we get a new image.  Tweak labels
so that everything lines up.