cool feature preview: streaming responses!
It's a little jerky, and a lot hacky, but it kinda works!
Pass in --stream-responses to have the bot's responses streamed
into a single message.

This works by continuously editing the bot's response message.
So you'll see an "edited" tag on the message.  But if you can
ignore that, that's cool.

Also, Discord seems to have some limits on how fast edits can be
made, so it waits at least 0.2 seconds in between update attempts.
It may wait longer than this, but it shouldn't be any faster.

I don't consider this quite baked yet, but I wanted to put it
out there and see what people thought.