add message_regex setting for dividing responses, inline stop tokens
message_regex:
Some newer chat-specific models are trained to generate specific
delimiters to separate their response into individual messages.

This adds a setting to tell the bot what regex to use to split
such responses into individual messages.

If this is set, it will only effect *how* message-splitting happens,
not whether it happens.  By default, the bot will still split messages.
But if "stream_responses" or "dont_split_responses" is enabled,
this setting will be ignored, as the messages won't be split anyway.

inline stop tokens:
with this change, we'll now look for stop tokens even if they're not
on a separate line.  The part of the message before the stop token
will be displayed, but anything after it will be hidden and the
response will be finished.

Also, automatically add anything from
  oobabot > request_params > stopping_strings
into the list of stop tokens to look for.