if a config setting had an incomplete dict, work rather than fail
If the config yaml had a dict parameter, and the user supplied
some but not all of the keys we would need, we would fail to
start.

Now we will add in default values for anything that the user
didn't supply.

The downside of this is that it's no longer possible to REMOVE
a setting entirely, if it's included in our default.  But this
seems like a less dangerous behavior than forcing the user to
supply every key.