limit the bot's ability to @-mention, focus replies for summons
---

small change, then the complex one...

small change:
don't allow the bot to @-mention anyone but the user who it's replying
to.  This is to prevent users from tricking the bot into pinging broad
groups, should the admin have granted them permission to do so.

@-mentions will still work when used via the /say command, which I am
presuming will be used by trusted users

----

A somewhat complicated but hopefully useful change to how message
history is filtered.

One problem users have seen is that in busy channels, the following
pattern happens:

 user1: @Bot how are you today?
 user2: what should I have for dinner?
 bot: you should have cake for dinner

The reason for this is that the bot has been seeing the full chat
history, up to the point that it generates the reply.  Even though
it was "summoned" by the first user, the AI didn't really know that,
and was just replying to the thread based on what it felt like saying.

With this change, when the bot was "summoned" (pinged with an @-mention
or with a wakeword), it will no longer see any chat history AFTER that
summon.

So in the above case, the bot would now only see:

 user1: @Bot how are you today?
 bot: [replying to User1's message]
      I'm feeling wonderful!

In addition, the bot will mark this message as an explicit reply to
the message that summoned it.

When the bot was not directly summoned, but instead is replying in an
"unsolicited" manner, it will see the full view of the history, and
its messages will not be marked as a reply.

For instance:

 user1: @Bot how are you today?
 user2: what should I have for dinner?
 bot: [replying to User1's message]
      I'm feeling wonderful!
 bot: how about tacos?

----
The tricky part about doing this was that, previously, the code was
using "message replies" as a hack to record data about something else
-- in particular, to indicate that a message was a UX message generated
as part of its image-generation feature, rather than a chat message.
Unless we continue to filter these out, the bot can easily fall out of character, repeating UX messages, rather than generate meaningful chat messages.

So, like all great hacks, this has switched from looking at the
"message replies" flag to now using the "suppress_embeds" flag.
This means that Discord won't generate embeds from bot messages,
which is probably a good thing.