replace <@___user_id___> with the user's display name in history
When presenting the message history to the bot, replace the
<@00000000> syntax with the user's display name instead.  This is
to help the bot understand the message history, as well as to have
it not "leak" the encoding into its response.

Also, fix a regression in the last commit where we would include
extra lins in the message history.  Ooops!