adds a /logs command which returns the most recent log messages
Unlike other commands, this can only be invoked by the bot's
owner.

The logs will be returned as an html file.

It's not the most convenient format for viewing the logs,
but the real reason this command exists is to test the
feature to export the logs in this manner.  It will be used
in the future by the oobabooga web interface plugin.