issue 33: add new "unsolicited_channel_cap" feature
Adds a new `unsolicited_channel_cap` option in
`discord` section of `config.yml`.  It does this:

FEATURE PREVIEW: Adds a limit to the number of channels
the bot will post unsolicited messages in at the same
time.  This is to prevent the bot from being too noisy
in large servers.

When set, only the most recent N channels the bot has
been summoned in will have a chance of receiving an
unsolicited message.  The bot will still respond to
@-mentions and wake words in any channel it can access.

Set to 0 to disable this feature.

see:
https://github.com/chrisrude/oobabot/issues/33