Fix #38 -- only reference the first message in a chain when replying to a message
Playing around with this, it seems that having it just reply-mention
the very first message in a chain works the best. The reason is that
the requests to oobabooga should be serialized anyway, so we will
always finish off the first reply in its entirety before starting
on an unsolicited reply.

Ironically, this is how things worked in the very first implementation. :)

Also, remove a test assert I mistakenly left in from the previous
commit.