add initial rust code for audio handling
This is a ton of work which was done on a seprate branch.

I don't care so much about the history being in this repo,
so I'm just going to commit it all at once.

Right now both projects just coexist in the same directory,
you can use poetry to build the python project, and cargo to
build the rust project.

Building the rust project creates a command-line app called
"disrivener" which will be able to connect to a discord voice
session, hear audio, and transcribe it to text.

At the moment one would use it by building both, launching
oobabot, joining a voice channel, and then looking for the
command-line message that oobabot will print out.  It will
be a command to invoke discrivener with a whole bunch of
arguments.  Just run that, and discrivener will connect.

Otherwise, the two parts don't talk to each other yet,
so the next step is to get them both building under maturin,
and then to create python stubs for discrivener using PyO3.