merge: big commit -- move to maturin build
This needs more documentation than I have time for right
now, but in short you need to install maturin and then
run `maturin develop`.  This should generate oobabot as
normal, but also generate a rust library and link it in!

The build isn't yet what I'd call "stable", and there are
some weird gotchas.  For example, maturin doesn't let
you run a clean build.  But you can have cargo run a
clean build for you, with
  `cargo clean --manifest-path ./rust/Cargo.toml`
but that only cleans the cargo side, and leaves the
*.so files that maturin puts in your PYTHON SOURCE DIRECTORIES.

Anyway, it works, so checking in for now, will fix it up later.