Update sample config.yml for version 0.2.0
Adds the following yaml settings:

 - stream_responses_speed_limit
When in "streaming" mode (i.e. when stream_responses is set to True), this will limit the
rate at which we update the streaming message in Discord.  We need this setting because Discord
has rate-limiting logic, and if we send updates "too fast" then it will slow down our updates
drastically, which will appear as jerky streaming.

This value is the minimum delay in seconds in between updates.  That is -- we will update Discord
no more than once this number of seconds.  The updates may come slower than this, perhaps on systems
that take a long time to generate tokens.  It's only guaranteed that they won't be any faster than this.

discrivener_location: and discrivener_model_location:

These are new settings to add voice support to oobabot.  Voice support means that the bot
can join voice chat channels, transcribe what is said, hear wakewords, and generate voice
responses in those channels.  All of the audio processing -- text to speech, and speech to
text -- is handled in a binary called "discrivener", whose source lives at
https://github.com/chrisrude/discriviner

I've tested this to work on Linux and OSX, but there is still more work to do in documenting
and packaging the software.  So for now, these settings are blank by default, which will
leave oobabot in text-only mode, as it has been.

command_lobotomize_response:

A user noticed that there was no setting to customize the text that gets shown when you use the
/lobotomize command.  Whoops!  Now here it is.  This is of particular interest because the bot
will see this text after a lobotomize occurs, so if you have specific character styling you want
to keep it from getting confused about, then you might want to put in custom text of your choosing
here.  You can also use variables {AI_NAME} and {USER_NAME} to represent the name of the AI, and
the name of the user who ran the /lobotomize command.