fix issue #63 -- history lost in streaming mode
In streaming mode, we edit previous messages as new tokens
arrive.  When we did this, we would drop the suppress_embeds=True
flag, which is our hack to tell whether a message should be
considered part of the history the bot sees (like most messages)
or not (like a message related to image generation, or a non-
character command response).

This is because the message.edit() method didn't support the
suppress_embeds= argument, so I left it out and hoped it would work.
Turn out it doesn't.  But... it also turns out it has a suppress=
argument, which does the same thing.  So use that!

After this fix, the bot still won't see streaming messages generated
by a buggy build, but new ones it generates should show up.

I really should have included this fix in 0.2.0, but
no worries.  It's in now.