Add new Stable Diffusion customizable_params setting.
This is a list of Stable Diffusion parameters that the user can
override from within the image request message itself.

My default, the following values may be overridden:
        "cfg_scale",
        "enable_hr",
        "model",
        "negative_prompt",
        "sampler_name",
        "seed",
        "height",
        "width",

If you're a super expert on the Stable Diffusion UI, and have
had lots of coffee today, you might have noticed that "model" isn't
actually a SD parameter.  Instead it's supposed to be:
 ['override_params']['sd_model_checkpoint']

We're going to call it 'model' instead, to make things easier for us.
If the user really wants to set the 'sd_model_checkpoint' they still
can, but otherwise this can be our secret!

The sampler and model will be checked against the list of models loaded
onto the SD server when oobabot launches.  This is to try and sanitize the
input somewhat, to keep users from doing nasty things.  It also means we
might not allow models that are loaded after we start running.  The user will
need to restart oobabot for us to be able to use them.

As a side-effect of this, you can now set default value for a sampler or
model in your config.yml file!  If you don't know the exact name, don't worry,
Oobabot will print a list of what it will allow on startup.  Just use that.

These parameters can be overridden by the Discord user
by including them in their image generation request.

The format for this is: param_name=value

This is a whitelist of parameters that can be overridden.
They must be simple parameters (strings, numbers, booleans),
and they must be in the request_params dictionary.

The value the user inputs will be checked against the type
from the request_params dictionary, and if it doesn't match,
the default value will be used instead.

Otherwise, this value will be passed through to Stable
Diffusion without any changes, so be mindful of what you allow
here.  It could potentially be used to inject malicious
values into your SD server.

For example, letting users set steps=1000000 could be bad for your server.