add oobabot.logs() and oobabot.log_count() methods
logs is just an alias for fancy_logger.recent_logs.get()
I wanted this as part of the oobabot class to make it clear
which methods are part of the public API and which are
internal.

log_count is a new method that returns the number of times
the log has been appended to. This is useful for an optimization
in the web UI, where we can check if the log has changed since
the last time it was fetched.