Unicode logging reliability fix ooba_client.py
Looks like oobabooga developer was the one who caused this Unicode issue. Refactored the code to enable the "bugfix" only in case it's actually needed. I don't trust developers who do everything on the main branch all the time not repeatedly cause this, so the bugfix is only applied when needed.