Address https://github.com/chrisrude/oobabot-plugin/issues/32
This adds a new setting: 'disable_unsolicited_replies'.  If set to
True, the bot will never reply to a message that does not @-mention
or a wakeword.

  # If set, the bot will not reply to any messages that do not @-mention it or include a
  # wakeword.  If unsolicited replies are disabled, the unsolicited_channel_cap setting will
  # have no effect.
  #   default: False
  disable_unsolicited_replies:

It might have been nice to reuse the "unsolicited_channel_cap" setting
for this purpose, but I had two concerns with that approach:

a. currently setting "unsolicited_channel_cap" to 0 means "no
   limit".  I didn't want to change that meaning in case people
    are using it that way already.

b. it's a bit of an indirect way to express the idea of "don't
   reply to unsolicited messages".

That said, the downside of this approach is that there are now
two settings which kinda interact with each other, but I think our
users can figure it out.